+++ test.py
@@ -1,0 +1,4 @@
+from platenumber import get_plate_number
+
+for i in range(1, 8):
+    print(get_plate_number('picture/' + str(i) + '.jpg'))
